+++ packages/amplify-sdk/src/amplify-sdk.js
@@ -114,7 +114,7 @@ export default class AmplifySDK {
 			 * @returns {Promise<Object>} Resolves the original account info object.
 			 */
 			findSession: async account => {
-				const result = await this.request('/api/v1/auth/findSession', account, {
+				let result = await this.request('/api/v1/auth/findSession', account, {
 					errorMsg: 'Failed to find session'
 				});
 				account.isPlatform = !!result;
@@ -160,6 +160,20 @@ export default class AmplifySDK {
 					organization: user.organization,
 					phone:        user.phone
 				});
+
+				result = await this.team.list(account, org.org_id);
+				if (result.teams.length) {
+					const team = result.teams.find(t => t.default) || { default: true, ...result.teams[0] };
+					account.team = {
+						default: team.default,
+						guid:    team.guid,
+						name:    team.name,
+						roles:   team.users?.find(u => u.guid === user.guid)?.roles || [],
+						tags:    team.tags
+					};
+				} else {
+					account.team = null;
+				}
 
 				return account;
 			},
@@ -855,6 +869,31 @@ export default class AmplifySDK {
 			}),
 
 			/**
+			 * Retrieves the list of environments associated to the user's org.
+			 * @param {Object} account - The account object.
+			 * @returns {Promise<Array>}
+			 */
+			environments: async account => {
+				assertPlatformAccount(account);
+				return await this.request('/api/v1/org/env', account, {
+					errorMsg: 'Failed to get organization environments'
+				});
+			},
+
+			/**
+			 * Retrieves the organization family used to determine the child orgs.
+			 * @param {Object} account - The account object.
+			 * @param {Object|String|Number} org - The organization object, name, id, or guid.
+			 * @returns {Promise<Object>}
+			 */
+			family: async (account, org) => {
+				const { id } = this.resolveOrg(account, org);
+				return await this.request(`/api/v1/org/${id}/family`, account, {
+					errorMsg: 'Failed to get organization family'
+				});
+			},
+
+			/**
 			 * Retrieves organization details for an account.
 			 * @param {Object} account - The account object.
 			 * @param {String} org - The organization object, name, id, or guid.
@@ -878,7 +917,7 @@ export default class AmplifySDK {
 
 				const { teams } = await this.team.list(account, id);
 
-				return {
+				const result = {
 					active:           org.active,
 					created:          org.created,
 					childOrgs:        null, // deprecated
@@ -896,31 +935,14 @@ export default class AmplifySDK {
 					userCount:        org.users.length,
 					userRoles:        org.users.find(u => u.guid === account.user.guid)?.roles || []
 				};
-			},
 
-			/**
-			 * Retrieves the list of environments associated to the user's org.
-			 * @param {Object} account - The account object.
-			 * @returns {Promise<Array>}
-			 */
-			environments: async account => {
-				assertPlatformAccount(account);
-				return await this.request('/api/v1/org/env', account, {
-					errorMsg: 'Failed to get organization environments'
-				});
-			},
+				if (org.entitlements?.partners) {
+					for (const partner of org.entitlements.partners) {
+						result[partner] = org[partner];
+					}
+				}
 
-			/**
-			 * Retrieves the organization family used to determine the child orgs.
-			 * @param {Object} account - The account object.
-			 * @param {Object|String|Number} org - The organization object, name, id, or guid.
-			 * @returns {Promise<Object>}
-			 */
-			family: async (account, org) => {
-				const { id } = this.resolveOrg(account, org);
-				return await this.request(`/api/v1/org/${id}/family`, account, {
-					errorMsg: 'Failed to get organization family'
-				});
+				return result;
 			},
 
 			/**
@@ -1147,12 +1169,12 @@ export default class AmplifySDK {
 
 				let org = params.org || account.org?.guid;
 				if (org) {
-					org = this.resolveOrg(account, params.org);
+					org = await this.org.find(account, org);
 					const { entitlements, subscriptions } = org;
 
 					roles = roles.filter(role => {
 						return role.org
-							&& (!role.partner || (entitlements.partners || []).includes(role.partner) && org[`${role.partner}.provisioned`])
+							&& (!role.partner || (entitlements.partners || []).includes(role.partner) && org[role.partner]?.provisioned)
 							&& (!role.entitlement || entitlements[role.entitlement])
 							&& (!role.subscription || subscriptions.find(sub => {
 								return new Date(sub.end_date) >= new Date() && role.subscription.includes(sub.product);